# 64-bit SIMD regs in ARM model, better BOUNDER_RULE, slow-ARM field optimizations