--- a/arm/allowed_asm
+++ b/arm/allowed_asm
@@ -41,4 +41,5 @@
 : subs$
 : tst$
 : umulh$
+: umull$
 : $