--- a/arm/proofs/bignum_ksqr_32_64_neon.ml
+++ b/arm/proofs/bignum_ksqr_32_64_neon.ml
@@ -917,7 +917,7 @@ let bignum_ksqr_32_64_neon_mc =
 let BIGNUM_KSQR_32_64_NEON_EXEC = ARM_MK_EXEC_RULE bignum_ksqr_32_64_neon_mc;;
 
 (* ------------------------------------------------------------------------- *)
-(* Proof for the inner-level 8->16 squring.                                  *)
+(* Proof for the inner-level 8->16 squaring.                                 *)
 (* ------------------------------------------------------------------------- *)
 
 let lemma1 = prove