@aqjune-aws 2023-11-17
I checked that the optimizations worked on my machine:
```
| (unit:ns)           | before | after | speedup |
| ------------------- | ------ | ----- | ------- |
| bignum_sqr_p256k1   | 37.1   | 32.5  | 14.15%  |
| bignum_montsqr_sm2  | 27.9   | 25.6  | 8.98%   |
| bignum_montsqr_p256 | 27.9   | 25.5  | 9.41%   |
```
And the updates to formal semantics and tactics did not cause significant slowdowns to running times of proof CIs.
The support for decoding ldr/str of 64-bit D registers looks good to me and `arm_ldstp_d` must be correct since it is a natural application of `arm_LDP`/`arm_STP` to the D registers.