@aqjune-aws 2023-11-17
I found that the green Merge button was now enabled - I will carefully exercise this greater power.
The changes are mostly to the formalization part (which was reviewed by me), and the remaining updates to the assembly codes are also formally verified. The updates don't introduce duplicated labels or use seemingly unfamiliar instructions.
Therefore, I believe that for this specific PR my review is enough. I will hit merge button. :)